requirements.txt
selenium
webdriver-manager
pandas
beautifulsoup4
python-dotenv
requests
